More refactor & cleanup
