+++ src/routes/class-dynamics-crm-route.php
@@ -107,26 +107,6 @@ class Dynamics_Crm_Route extends Base_Route {
     ];
   }
 
-
-  /**
-   * Removes some params we don't want to send to CRM from request.
-   *
-   * @param  array $params Params received in request.
-   * @return array
-   */
-  protected function unset_irrelevant_params( array $params ): array {
-    $filtered_params   = [];
-    $irrelevant_params = array_flip( $this->get_irrelevant_params() );
-
-    foreach ( $params as $key => $param ) {
-      if ( ! isset( $irrelevant_params [ $key ] ) ) {
-        $filtered_params[ $key ] = $param;
-      }
-    }
-
-    return $filtered_params;
-  }
-
   /**
    * Defines a list of required parameters which must be present in the request or it will error out.
    *
